requirements.txt
pycketcasts
coloredlogs
python-dotenv
mygpoclient